moved layers implementation to src
