+++ libs/ml/include/ml/layers/PRelu.hpp
@@ -17,16 +17,8 @@
 //
 //------------------------------------------------------------------------------
 
-#include "math/tensor/tensor.hpp"
 #include "ml/layers/fully_connected.hpp"
 #include "ml/ops/prelu_op.hpp"
-#include "vectorise/fixed_point/fixed_point.hpp"
-
-#include <cmath>
-#include <cstdint>
-#include <random>
-#include <string>
-#include <vector>
 
 namespace fetch {
 namespace ml {
@@ -46,42 +38,9 @@ public:
   PRelu() = default;
 
   explicit PRelu(uint64_t in, std::string const &name = "PRelu",
-                 WeightsInit init_mode = WeightsInit::XAVIER_GLOROT)
-  {
-    std::string input =
-        this->template AddNode<fetch::ml::ops::PlaceHolder<TensorType>>(name + "_Input", {});
+                 WeightsInit init_mode = WeightsInit::XAVIER_GLOROT);
 
-    std::string alpha =
-        this->template AddNode<fetch::ml::ops::Weights<TensorType>>(name + "_Alpha", {});
-
-    TensorType alpha_data(std::vector<SizeType>({in, 1}));
-    fetch::ml::ops::Weights<TensorType>::Initialise(alpha_data, in, in, init_mode);
-
-    this->SetInput(alpha, alpha_data);
-
-    std::string output = this->template AddNode<fetch::ml::ops::PReluOp<TensorType>>(
-        name + "_PReluOp", {input, alpha});
-
-    this->AddInputNode(input);
-    this->SetOutputNode(output);
-
-    this->Compile();
-  }
-
-  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override
-  {
-    // get base class saveable params
-    std::shared_ptr<OpsSaveableParams> sgsp = SubGraph<TensorType>::GetOpSaveableParams();
-
-    auto ret = std::make_shared<SPType>();
-
-    // copy subgraph saveable params over
-    auto sg_ptr1 = std::dynamic_pointer_cast<typename SubGraph<TensorType>::SPType>(sgsp);
-    auto sg_ptr2 = std::dynamic_pointer_cast<typename SubGraph<TensorType>::SPType>(ret);
-    *sg_ptr2     = *sg_ptr1;
-
-    return ret;
-  }
+  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override;
 
   void SetOpSaveableParams(SPType const &sp)
   {
