+++ libs/ml/src/ops/activation.cpp
@@ -65,7 +65,30 @@ std::string AddActivationNode(ActivationType type, Graph<T> *g, std::string name
 ///////////////////////////////
 /// EXPLICIT INSTANTIATIONS ///
 ///////////////////////////////
-
+template std::string AddActivationNode<math::Tensor<int8_t>>(ActivationType               type,
+                                                             Graph<math::Tensor<int8_t>> *g,
+                                                             std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<int16_t>>(ActivationType                type,
+                                                              Graph<math::Tensor<int16_t>> *g,
+                                                              std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<int32_t>>(ActivationType                type,
+                                                              Graph<math::Tensor<int32_t>> *g,
+                                                              std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<int64_t>>(ActivationType                type,
+                                                              Graph<math::Tensor<int64_t>> *g,
+                                                              std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<uint8_t>>(ActivationType                type,
+                                                              Graph<math::Tensor<uint8_t>> *g,
+                                                              std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<uint16_t>>(ActivationType                 type,
+                                                               Graph<math::Tensor<uint16_t>> *g,
+                                                               std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<uint32_t>>(ActivationType                 type,
+                                                               Graph<math::Tensor<uint32_t>> *g,
+                                                               std::string name, std::string input);
+template std::string AddActivationNode<math::Tensor<uint64_t>>(ActivationType                 type,
+                                                               Graph<math::Tensor<uint64_t>> *g,
+                                                               std::string name, std::string input);
 template std::string AddActivationNode<math::Tensor<float>>(ActivationType              type,
                                                             Graph<math::Tensor<float>> *g,
                                                             std::string name, std::string input);
